source/backend/requirements.txt
openai
Python-dotenv